Added the download function to perform dssp annotation and mask the sequence files in an incremental way.

+++ idr-lcr-struct/ssDssp.py
@@ -15,7 +15,7 @@ import pandas as pd
 import numpy as np
 import math
 import re
-import gzip, shutil, csv, pickle, shutil, time
+import gzip, shutil, pickle, time
 
 import os, subprocess, shlex
 
@@ -60,7 +60,7 @@ def extract_seqres(path_seqres):
     return seqres_data, seqres_keys, seqres_desc, seqres_fnb
 
 
-def get_missing_names(seqres_keys, path_masked=""):
+def get_missing_names(seqres_keys, path_masked="", sep="_"):
     ''' Extract the list of files not available on the last extraction.
     If no path_masked is provided we assume this is the first execution, so all
     protein PDB files will be downloaded.
@@ -68,37 +68,77 @@ def get_missing_names(seqres_keys, path_masked=""):
     if (path_masked!=""):
         ss_dict = resources.load_seqs(path_masked, "|", 'pdb')
         old_ss_keys = np.array(list(ss_dict.keys()))
-        missing_ids = np.setdiff1d(seqres_keys, old_ss_keys)
-        obsolete_ids = np.setdiff1d(old_ss_keys, seqres_keys)
-    return missing_ids, obsolete_ids
+        missing_names = np.setdiff1d(seqres_keys, old_ss_keys)
+        missing_ids = np.unique([n.split(sep)[0] for n in missing_names])
+        obsolete_names = np.setdiff1d(old_ss_keys, seqres_keys)
+    return missing_names, missing_ids, obsolete_names
 
 
-def select_target(path_selected, un_pdb_ids, sep="_"):
+def select_target(path_selected, un_pdb_ids, intersect=True, sep="_"):
     ''' Gets all files and extract just the target list based on an external 
     list of IDs. 
-            un_pdb_ids must have the PDB ID and chain separated by sep or just
-            the PDB ID with no chain.
+            un_pdb_ids must have the PDB ID WHITOUT CHAIN.
     '''
     files = sorted([f.path for f in os.scandir(path_selected) if (re.search(r'\W*.gz', f.path))])
     base_files = [os.path.splitext(os.path.splitext(os.path.basename(f))[0])[0] for f in files]
-    un_ids = np.unique([n.split(sep)[0] for n in un_pdb_ids])
     sel_files = []
-    for x, y in zip(base_files, files):
-        if (x in un_ids):
-            sel_files.append(y)
-    return(sel_files)
+    if (intersect):
+        for x, y in zip(base_files, files):
+            if (x in un_pdb_ids):
+                sel_files.append(y)
+    else:
+        sel_files = np.setdiff1d(un_pdb_ids, base_files)
+    return sel_files
 
 
-def download_missing(path_pdb_files, missing_ids):
-    un_ids = np.unique([n.split(sep)[0] for n in un_pdb_ids])
+def manage_missing(path_pdb_files, missing_ids, file_name):
+    ''' Deletes old PDB/CIF files that showed errors previously so the script
+    can try to run DSSP again and download the .pdb file using the bash file 
+    provided by PDB and when not available the .cif file. Every 10 executions 
+    it checks if it should abort in case some files remain missing.
+    '''
+    
+    # Removing files already existing in the folder (There may be errors from 
+    # last execution).
     sel_files = select_target(path_pdb_files, list(missing_ids))
-    try:
-        for file in sel_files:
-            os.remove(file)
-    except Exception as e:
-        print([base, e])
+    if len(sel_files)>0 :
+        try:
+            for file in sel_files:
+                os.remove(file)
+        except Exception as e:
+            print([base, e])
+    
+    # Now download all missing .pdb.gz files from PDB
+    source_sh = os.path.join(path_pdb_files, "batch_download.sh")
+    # Save missing files to disk, so the .sh provided by PDB can run over all 
+    # downloads at once.
+    file_name = os.path.join(path_pdb_files, "missing_pdbs.txt")
+    missing = list(missing_ids)
+    resources.save_sep_file(missing, file_name)
+    prev_miss = len(missing)
+    file_type = "-p"
+    i = 0
+    while len(missing) > 0:
+        i += 1
+        subprocess.call(shlex.split(f"{source_sh} -f {file_name} -o {path_pdb_files} {file_type}"))
+        missing = list(select_target(path_pdb_files, missing, False))
+        resources.save_sep_file(missing, file_name)
+        if len(missing)!=prev_miss:
+            prev_miss = len(missing)
+        else:
+            file_type = "-c"
+        if (i%10==0 and file_type=="-c" and len(missing) > 0:
+            print()
+            print("{0} Attemps were already made and there are still missing PDB/CIF files.\nPlease download them manually.".format(str(i)))
+            break
+    os.remove(file_name)
+    return missing
+        
+    
+    
 
 
-def main():
+def main(comp_path):
     seqres_data, seqres_keys, seqres_desc, seqres_fnb = extract_seqres(path_seqres)
-    missing_ids, obsolete_ids = get_missing_names(seqres_keys, path_masked)
+    missing_names, missing_ids, obsolete_names = get_missing_names(seqres_keys, path_masked)
+
